% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotatedMM.R
\docType{package}
\name{annotatedMM}
\alias{annotatedMM}
\alias{annotatedMM-package}
\title{annotatedMM: A package for reading and writing "annotated" Matrix Market (AMM) files}
\description{
The main functions of interest are writeAMM and readAMM.
}
