% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCommentLineMM.R
\name{getCommentLineMM}
\alias{getCommentLineMM}
\title{Return 1:N comment lines from a Matrix Market file}
\usage{
getCommentLineMM(readConn, nLines = 1)
}
\arguments{
\item{readConn}{an open file connection}

\item{nLines}{optional number of lines to read, default 1}
}
\value{
A character vector of comment line/s starting from current file position
}
\description{
Return 1:N comment lines from a Matrix Market file
}
