% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMetadataName.R
\name{getMetadataName}
\alias{getMetadataName}
\title{Return metadata key name from input metadata line}
\usage{
getMetadataName(line = NULL)
}
\arguments{
\item{line}{Metadata line}
}
\value{
A string corresponding to the metadata key
}
\description{
Return metadata key name from input metadata line
}
