% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMetadataValues.R
\name{getMetadataValues}
\alias{getMetadataValues}
\title{Return metadata values as vector of tokens}
\usage{
getMetadataValues(line = NULL)
}
\arguments{
\item{line}{Metadata line}
}
\value{
A string corresponding to the metadata key
}
\description{
Return metadata values as vector of tokens
}
