% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isMetadataLine.R
\name{isMetadataLine}
\alias{isMetadataLine}
\title{Return whether or not a line from file looks like a metadata line}
\usage{
isMetadataLine(line = NULL)
}
\arguments{
\item{line}{Line from file}
}
\value{
Boolean TRUE or FALSE
}
\description{
Return whether or not a line from file looks like a metadata line
}
