% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readAMM.R
\name{readAMM}
\alias{readAMM}
\title{Read an "Annotated" Matrix Market exchange file and return a sparseMatrix}
\usage{
readAMM(fileName = NULL)
}
\arguments{
\item{fileName}{name of Matrix Market file to read}
}
\value{
A sparseMatrix (or descendant) object of type returned by Matrix::readMM
}
\description{
Reads an "Annotated" Matrix Market file and returns a corresponding
sparseMatrix (or descendant e.g. dgCMatrix). If row and column names are
found they will be attached to the returned sparse matrix object.
If reading a normal Matrix Market exchange file, this function should
return result identical to Matrix::readMM.
}
