% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeAMM.R
\name{writeAMM}
\alias{writeAMM}
\title{Write an "Annotated" Matrix Market exchange file}
\usage{
writeAMM(object = NULL, fileName = NULL)
}
\arguments{
\item{object}{sparse matrix object to write to file, any type supported by Matrix::writeMM}

\item{fileName}{name of file to create (will add "amtx" extension if no extension given)}
}
\value{
TRUE on success, FALSE otherwise
}
\description{
TODO: Ensure documentation is accurate
Writes a sparseMatrix (e.g. dgCMatrix object) to an "Annotated" Matrix Market
file. Row and column names are written to the output "AMTX" file. An AMTX file
is still a valid MTX file so the output should interroperate with other tools
albeit in a reduced capacity, for instance row and column names would not
be parsed as they are stored in the comments section of the file. The numeric
content of the file should always be preserved however. NOTE: The Matrix
Market exchange format is variously referred to as "MM", "MTX" or "MEX".
}
\examples{
writeAMM()

}
