% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readAMM.R
\name{readAMM}
\alias{readAMM}
\title{Read an "Annotated" Matrix Market exchange file and return a sparse matrix}
\usage{
readAMM(fileName = NULL)
}
\arguments{
\item{fileName}{name of file to read}
}
\value{
A sparse matrix object of type returned by Matrix::readMM
}
\description{
TODO: Ensure documentation is accurate
Reads an "Annotated" Matrix Market file. If row and column names were 
recorded in the file, they will be attached to the returned sparse matrix
object.
}
\examples{
readAMM()

}
