% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeAMM.R
\name{writeAMM}
\alias{writeAMM}
\title{Write an "Annotated" Matrix Market exchange file}
\usage{
writeAMM(object = NULL, fileName = NULL)
}
\arguments{
\item{object}{sparseMatrix object to write to file}

\item{fileName}{name of file to create}
}
\value{
TRUE on success, FALSE otherwise
}
\description{
Writes a sparseMatrix (e.g. dgCMatrix object) to an "Annotated" Matrix
Market file ("AMM"). Row and column names are written to the output AMM
file. An AMM file is still a valid MM file so the output should inter-
operate with other tools albeit in a reduced capacity, for instance row
and column names would not be parsed as they are stored in the comments
section of the file. The numeric content of the file should always be
preserved, however. NOTE: The Matrix Market exchange format is variously
known as "MM", "MTX" or "MEX".
}
